+++ cmd/hidden_lake/applications/messenger/internal/config/types.go
@@ -25,7 +25,6 @@ type IConfig interface {
 type IConfigSettings interface {
 	GetMessagesCapacity() uint64
 	GetWorkSizeBits() uint64
-	GetShareEnabled() bool
 	GetPseudonym() string
 	GetStorageKey() string
 	GetLanguage() language.ILanguage